% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_frequency.R
\name{age_frequency}
\alias{age_frequency}
\title{Plots distribution of ages across survey years.}
\usage{
age_frequency(
  data,
  subregions = c("AK ALEUTIANS", "AK BERING", "AK GULF", "NWFSC", "PBS"),
  species,
  by_sex = FALSE,
  cutoff = 0.95,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing age and sex information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, AK ALEUTIANS, and/or AK BERING. Default all.}

\item{species}{species common or scientific name.}

\item{by_sex}{TRUE or FALSE for if you want to differentiate by sex.}

\item{cutoff}{define a cutoff for grouping older ages together.}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a ggplot object
}
\description{
Plots distribution of ages across survey years.
}
