% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.dbi.R
\docType{data}
\name{all.dbi}
\alias{all.dbi}
\title{Dataset of design based biomass indices for all science centers}
\format{
\subsection{\code{all.dbi}}{

A data frame with 4154 rows and 8 columns:
\describe{
\item{region}{Research center that gathered data}
\item{survey_group}{Grouping of surveys used for shiny app}
\item{survey}{survey where data was collected}
\item{common_name}{Common name of specimen collected }
\item{scientific_name}{Scientific name of specimen collected}
\item{year}{Year specimen was collected}
\item{est}{Estimated design-based biomass}
\item{lwr}{Lower confidence interval of biomass estimation}
\item{upr}{Upper confidence interval of biomass estimation}
...
}
}
}
\usage{
all.dbi
}
\description{
Joined data from design based biomass estimates calculated separately by each science center
}
\keyword{datasets}
