% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data_document.R
\docType{data}
\name{all_catch}
\alias{all_catch}
\title{all_catch: Catch data (occurrences) for all surveys/regions}
\format{
A data frame with 3708 rows and 8 variables:
\describe{
\item{region}{region of data collection}
\item{survey}{subregion}
\item{common_name}{common name of specimen}
\item{scientific_name}{scientific name of specimen}
\item{year}{year tows were conducted}
\item{n_tows}{total number of tows conducted}
\item{n_pos}{number of positive tows}
\item{proportion_pos}{proportion of total tows that were positive}

}
}
\usage{
all_catch
}
\description{
Contains catch data from all surveys/regions for available top species. Catch data pulled from packages nwfscSurvey and surveyjoin
}
\keyword{datasets}
