% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbi.R
\docType{data}
\name{catch}
\alias{catch}
\title{Main function to display survey specimen counts}
\format{
An object of class \code{data.frame} with 13338 rows and 26 columns.
}
\usage{
catch
}
\arguments{
\item{data}{a data frame with year, estimated biomass, lower and upper SE}
}
\value{
a ggplot object
}
\description{
Main function to display survey specimen counts
}
\examples{
\dontrun{
plot_dbi(biomass.data)
}
}
\keyword{datasets}
