% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_plot.R
\name{depth_plot}
\alias{depth_plot}
\title{Function for age-depth plot}
\usage{
depth_plot(
  data,
  subregion = c("NWFSC", "PBS", "AK BSAI", "AK GULF"),
  common,
  sex = FALSE
)
}
\arguments{
\item{data}{biological data}

\item{subregion}{choose PBS, NWFSC, AK BSAI, AK GULF}

\item{common}{species common name}

\item{sex}{show sex differentiation}
}
\value{
a ggplot object
}
\description{
Function for age-depth plot
}
\examples{
\dontrun{
all_data <- bind_rows(afsc_bio, pbs_bio, nwfsc_bio)
depth_plot(all_data, "AFSC", "arrowtooth flounder")
}
}
