% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmap.R
\name{fishmap}
\alias{fishmap}
\title{Function for modeled CPUE map based on prediction data}
\usage{
fishmap(
  data,
  subregion = c("AFSC", "NWFSC", "PBS", "AK BSAI", "AK GULF"),
  common_name
)
}
\arguments{
\item{data}{prediction data from fishfit scripts}

\item{subregion}{choose AK BSAI, AK GULF, PBS, NWFSC}

\item{common_name}{species common name}
}
\value{
a ggplot object
}
\description{
Function for modeled CPUE map based on prediction data
}
\examples{
\dontrun{
data(predictions_afsc)
data(predictions_pbs)
data(predictions_nwfsc)
data <- bind_rows(predictions_afsc, predictions_pbs, predictions_nwfsc)

fishmap(data, c("AFSC", "PBS", "NWFSC"), "arrowtooth flounder")
fishmap(data, c("AK BSAI", "NWFSC"), "sablefish")
fishmap(data, "PBS", "dover sole")
}
}
