% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{length_frequency}
\alias{length_frequency}
\title{Function to plot length frequency of fish species}
\usage{
length_frequency(data, region, time_series = TRUE)
}
\arguments{
\item{data}{biological fisheries data containing length information}

\item{region}{choose a science center (AFSC, NWFSC, PBS)}

\item{time_series}{TRUE or FALSE}
}
\value{
a ggplot object
}
\description{
Function to plot length frequency of fish species
}
\examples{
\dontrun{
data(nwfsc_bio)
data(afsc_bio)
data(pbs_bio)
all_data <- rbind(nwfsc_bio, afsc_bio, pbs_bio)
data <- all_data |> filter(common_name == "arrowtooth flounder")

length_frequency(data, region = c("AFSC", "PBS", "NWFSC"))

}
}
