% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_ts.R
\name{length_ts}
\alias{length_ts}
\title{Main function to plot mean lengths over time by sex}
\usage{
length_ts(
  data,
  subregions = c("AK BSAI", "AK GULF", "PBS", "NWFSC"),
  species,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing length information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name.}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted.#' @return a time series plot of mean lengths by sex}
}
\description{
Main function to plot mean lengths over time by sex
}
\examples{
\dontrun{
data(pbs_bio)
data(afsc_bio)
data(nwfsc_bio)
all_data <- bind_rows(pbs_bio, afsc_bio, nwfsc_bio)

length_ts(all_data, species = "arrowtooth flounder")
length_ts(all_data, c("NWFSC", "AK GULF"),species = "anoplopoma fimbria", facet_all = F)
}
}
