% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{length_weight}
\alias{length_weight}
\title{Main function to plot length-weight relationships by sex}
\usage{
length_weight(
  data,
  subregions = c("AK BSAI", "AK GULF", "PBS", "NWFSC"),
  species,
  subset = TRUE,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing age, length, and sex information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name.}

\item{subset}{default TRUE for a faster plotting subset of n = 10000. Set FALSE for all available data.}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted.}
}
\value{
a plot of sexed data with log-log regression slope and intercept
}
\description{
Main function to plot length-weight relationships by sex
}
\examples{
\dontrun{
data(vb_predictions)
data(pbs_bio)
data(afsc_bio)
data(nwfsc_bio)
all_data <- bind_rows(pbs_bio, afsc_bio, nwfsc_bio)

length_weight(all_data, species = "arrowtooth flounder")
length_weight(all_data, c("NWFSC", "AK GULF"),species = "anoplopoma fimbria", facet_all = F)
}
}
