% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cpue.R
\name{map_cpue}
\alias{map_cpue}
\title{Function for CPUE map for US coast data}
\usage{
map_cpue(data, map_type)
}
\arguments{
\item{data}{catch data from pull_catch(), containing cpue and long/lat coordinates}

\item{map_type}{choose either "static" or "dynamic"}
}
\value{
a ggplot object
}
\description{
Function for CPUE map for US coast data
}
\examples{
\dontrun{
map_cpue(catch_data, map_type = "static")
}
}
