% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{my_func}
\alias{my_func}
\title{Main function to summarize relationship between x and y}
\usage{
my_func(data)
}
\arguments{
\item{data}{a data frame with columns x and y}
}
\value{
a data frame of joined haul and catch data
}
\description{
Main function to summarize relationship between x and y
}
\examples{
\dontrun{
set.seed(123)
df <- data.frame(x = 1:10, y = rnorm(10))
result <- my_func(df)
}
}
