% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data_document.R
\docType{data}
\name{nwfsc_bio}
\alias{nwfsc_bio}
\title{nwfsc_bio: Biological data for NWFSC survey}
\format{
A data frame with 1491160 rows and 11 variables:
\describe{
\item{region}{region of data collection}
\item{survey}{subregion}
\item{year}{year specimen was collected}
\item{common_name}{common name of specien}
\item{scientific_name}{scientific name of specimen}
\item{sex}{specimen sex, M, F, or U (unknown)}
\item{length_cm}{length of specimen, in cm}
\item{weight_kg}{weight of specimen, in kg}
\item{age_years}{age of specimen, in years}
\item{otosag_id}{Unique to nwfsc dataset, defines number of available age structures}
\item{depth_m}{Depth in meters}
}
}
\usage{
nwfsc_bio
}
\description{
Contains biological data from NWFSC surveys for available top species defined in data-raw/nwfsc_top.rds.
}
\keyword{datasets}
