% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_depth.R
\name{plot_age_depth}
\alias{plot_age_depth}
\title{Function for age-depth plot}
\usage{
plot_age_depth(
  data,
  subregion = c("NWFSC", "PBS", "AK BSAI", "AK GULF"),
  common,
  by_sex = FALSE
)
}
\arguments{
\item{data}{biological data}

\item{subregion}{choose PBS, NWFSC, AK BSAI, AK GULF}

\item{common}{species common name}

\item{by_sex}{show sex differentiation}
}
\value{
a ggplot object
}
\description{
Function for age-depth plot
}
\examples{
\dontrun{
data(afsc_bio)
data(pbs_bio)
data(nwfsc_bio)
all_data <- bind_rows(afsc_bio, pbs_bio, nwfsc_bio)
plot_age_depth(all_data, c("AK BSAI", "AK GULF", "PBS", "NWFSC"), "arrowtooth flounder")
}
}
