% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbi.R
\name{plot_dbi}
\alias{plot_dbi}
\title{Function and formatted data to plot design based index}
\usage{
plot_dbi(species, reg)
}
\arguments{
\item{species}{common name of species of interest. See examples for options per dataset.}

\item{reg}{region, required to specify region to be plotted. See examples for options per dataset.}
}
\value{
a ggplot object, plot of design based indicies
}
\description{
Function and formatted data to plot design based index
}
\examples{
\dontrun{
# To see options for species and region in science centers: "NWFSC", "AFSC", "PBS"
data(all.dbi)

all.dbi \%>\%
filter(science_center == "AFSC") \%>\%
 distinct(common_name)

all.dbi \%>\%
filter(science_center == "AFSC") \%>\%
distinct(region)


# usage examples
plot_dbi("sablefish", "U.S. West Coast")
plot_dbi("Dover sole", "U.S. Gulf of Alaska")
plot_dbi("arrowtooth flounder", "SYN QCS")
}
}
