% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{plot von Bertalanffy function predictions}
\usage{
plot_growth(data, predictions, subregions, common, facet_all = TRUE)
}
\arguments{
\item{data}{biological data containing age, length, and sex information}

\item{predictions}{von Bertalanffy prediction data in /data folder}

\item{subregions}{choose NWFSC, PBS, AK Gulf, and/or AK BSAI}

\item{common}{species common name}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a ggplot object
}
\description{
plot von Bertalanffy function predictions
}
\examples{
\dontrun{
data(vb_predictions)
data(pbs_bio)
data(afsc_bio)
data(nwfsc_bio)
data <- bind_rows(pbs_bio, afsc_bio, nwfsc_bio)

plot_growth(data, vb_predictions, "AFSC", "walleye pollock")
plot_growth(data, vb_predictions, c("AK BSAI", "AK GULF", "NWFSC", "PBS"), "arrowtooth flounder")
}
}
