% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_dbi.R
\name{plot_stan_dbi}
\alias{plot_stan_dbi}
\title{Function and formatted data to plot standardized design based index}
\usage{
plot_stan_dbi(species, regions)
}
\arguments{
\item{species}{common name of species of interest. See unique(data$common_name) for options per dataset.}

\item{regions}{center or subregions to be plotted. Takes a character list of region names OR a single center name ( "NWFSC", "AFSC", or "PBS") . See unique(data$region) for subregion options per center dataset.}
}
\value{
a ggplot object, plot of standardized design based indicies of 1+ surveys.
}
\description{
Function and formatted data to plot standardized design based index
}
\examples{
\dontrun{
plot_stan_dbi("sablefish", "PBS")
plot_stan_dbi("sablefish", c("U.S. West Coast", "U.S. Gulf of Alaska"))
}
}
