% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_document.R
\docType{data}
\name{predictions_afsc}
\alias{predictions_afsc}
\title{Modeled prediction data for AFSC region}
\format{
A data frame with 2809818 rows and 10 variables:
\describe{
\item{lon}{Original longitude coordinate}
\item{lat}{Original latitude coordinate}
\item{X}{New UTM coordinate}
\item{Y}{New UTM coordinate}
\item{prediction}{Model estimate}
\item{species}{Species specification}
\item{sanity}{Check if the model/prediction has any issues}
\item{region}{Region AFSC}
\item{crs}{UTM crs used in coordinate transformation}
\item{survey}{Subregion}
}
}
\usage{
predictions_afsc
}
\description{
Contains prediction data from spatial models on AFSC surveys for select species.
}
\keyword{datasets}
