% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_document.R
\docType{data}
\name{predictions_pbs}
\alias{predictions_pbs}
\title{Modeled prediction data for PBS region}
\format{
A data frame with 321480 rows and 9 variables:
\describe{
\item{lon}{Original longitude coordinate}
\item{lat}{Original latitude coordinate}
\item{X}{New UTM coordinate}
\item{Y}{New UTM coordinate}
\item{prediction}{Model estimate}
\item{species}{Species specification}
\item{sanity}{Check if the model/prediction has any issues}
\item{region}{Region PBS}
\item{crs}{UTM crs used in coordinate transformation}
}
}
\usage{
predictions_pbs
}
\description{
Contains prediction data from spatial models on NWFSC surveys for select species.
}
\keyword{datasets}
