% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_vb_fit_doc.R
\docType{data}
\name{vb_predictions}
\alias{vb_predictions}
\title{Alternate data storage for von Bertalanffy predictions}
\format{
A data frame with 3061 rows and 9 variables:
\describe{
\item{age_years}{Fish age in years}
\item{fit}{Growth estimate (length cm)}
\item{sex}{Fish sex, male or female}
\item{linf}{Asymptotic length of growth curve}
\item{k}{growth coefficient, rate of growth}
\item{t0}{exponent, slope of length-weight regression }
\item{center}{Region (NWFSC, AFSC, PBS)}
\item{common}{Species common name}
\item{scientific}{Species scientific name}
}
}
\usage{
vb_predictions
}
\description{
Contains von Bertalanffy growth curve estimates for each fish age bracket
}
\keyword{datasets}
