% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonb_growth.R
\name{vonb_growth}
\alias{vonb_growth}
\title{von Bertalanffy growth function plot}
\usage{
vonb_growth(data)
}
\arguments{
\item{data}{catch data from pull_bio(), containing Length_cm, Age_years, and Sex}
}
\value{
a ggplot object
}
\description{
von Bertalanffy growth function plot
}
\examples{
\dontrun{
growth(bio_data)
}
}
