% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_frequency.R
\name{age_frequency}
\alias{age_frequency}
\title{Plots distribution of ages across survey years.}
\usage{
age_frequency(data, region, species, sex = FALSE, cutoff = 0.95)
}
\arguments{
\item{data}{biological fisheries data containing age and sex information}

\item{region}{choose science center NWFSC, AFSC, PBS}

\item{species}{species common name}

\item{sex}{TRUE or FALSE for if you want to differentiate by sex}

\item{cutoff}{define a cutoff for grouping older ages together}
}
\value{
a ggplot object
}
\description{
Plots distribution of ages across survey years.
}
\examples{
\dontrun{
data("nwfsc_bio")
data("afsc_bio")
data("pbs_bio")
nwfsc_bio <- nwfsc_bio |> select(-otosag_id)
all_data <- rbind(afsc_bio, nwfsc_bio, pbs_bio)

age_frequency(all_data, region = c("PBS", "NWFSC"), species = "yellowtail rockfish")
}
}
