% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_frequency.R
\name{age_frequency}
\alias{age_frequency}
\title{Plots distribution of ages across survey years.}
\usage{
age_frequency(data, sex = FALSE, cutoff = 0.95)
}
\arguments{
\item{data}{biological data from pull_bio(), containing Age_years}

\item{sex}{TRUE or FALSE for if you want to differentiate by sex}

\item{cutoff}{define a cutoff for grouping older ages together}
}
\value{
a ggplot object
}
\description{
Plots distribution of ages across survey years.
}
\examples{
\dontrun{
data <- pull_bio(survey = "NWFSC.Combo", common_name = "Pacific hake")

# Normal call
age_frequency(data)

# Include sex differentiation
age_frequency(data, sex = TRUE)

# Specify a cutoff for binning
age_frequency(data, cutoff = 0.85)
}
}
