% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_frequency.R
\name{age_frequency}
\alias{age_frequency}
\title{Plots distribution of ages across survey years.}
\usage{
age_frequency(
  data,
  subregions,
  common,
  sex = FALSE,
  cutoff = 0.95,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological fisheries data containing age and sex information}

\item{subregions}{choose NWFSC, PBS, AK Gulf, and/or AK BSAI}

\item{common}{species common name}

\item{sex}{TRUE or FALSE for if you want to differentiate by sex}

\item{cutoff}{define a cutoff for grouping older ages together}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a ggplot object
}
\description{
Plots distribution of ages across survey years.
}
\examples{
\dontrun{
data("nwfsc_bio")
data("afsc_bio")
data("pbs_bio")
all_data <- bind_rows(afsc_bio, nwfsc_bio, pbs_bio)

age_frequency(all_data, c("PBS", "NWFSC", "AK BSAI", "AK GULF"), "yellowtail rockfish")
}
}
