% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_frequency.R
\name{age_frequency}
\alias{age_frequency}
\title{Plots distribution of ages across survey years.}
\usage{
age_frequency(data, sex = FALSE, cutoff = 0.95)
}
\arguments{
\item{data}{biological fisheries data containing age and sex information}

\item{sex}{TRUE or FALSE for if you want to differentiate by sex}

\item{cutoff}{define a cutoff for grouping older ages together}
}
\value{
a ggplot object
}
\description{
Plots distribution of ages across survey years.
}
\examples{
\dontrun{
# US West Coast
load("data/nwfsc_bio.rda")
data <- nwfsc_bio |> filter(common_name == "arrowtooth flounder")
age_frequency(data)

# Canada
load("data/pbs_bio.rda")
data <- pbs_bio |> filter(common_name == "arrowtooth flounder")
age_frequency(data, sex = TRUE)

# Alaska 
load("data/afsc_bio.rda")
data <- afsc_bio |> filter(common_name == "arrowtooth flounder")
age_frequency(data, cutoff = 0.75)
}
}
