% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_plot.R
\name{depth_plot}
\alias{depth_plot}
\title{Function for age-depth plot}
\usage{
depth_plot(
  data,
  subregion = c("AFSC", "NWFSC", "PBS", "AK BSAI", "AK GULF"),
  common,
  option = c("a", "b", "c")
)
}
\arguments{
\item{data}{prediction data from fishfit scripts}

\item{region}{choose AFSC, PBS, NWFSC (choosing AK BSAI and AK GULF will default to AFSC)}

\item{common_name}{species common name}
}
\value{
a ggplot object
}
\description{
Function for age-depth plot
}
\examples{
\dontrun{
# Still need depth column in the regular _bio data frames
# Once that is added this code will work

all_data <- bind_rows(afsc_bio, pbs_bio, nwfsc_bio)
depth_plot(all_data, "AFSC", "arrowtooth flounder")
}
}
