% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmap.R
\name{fishmap}
\alias{fishmap}
\title{Function for modeled CPUE map based on prediction data}
\usage{
fishmap(data, common_name)
}
\arguments{
\item{data}{prediction data from fishfit scripts}

\item{common_name}{species common name}
}
\value{
a ggplot object
}
\description{
Function for modeled CPUE map based on prediction data
}
\examples{
\dontrun{
data(predictions_afsc)
fishmap(predictions_afsc, "arrowtooth flounder")
}
}
