% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{length_frequency}
\alias{length_frequency}
\title{Function to plot length frequency of fish species}
\usage{
length_frequency(
  data,
  subregions = c("AK BSAI", "AK GULF", "NWFSC", "PBS"),
  species,
  time_series = TRUE,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing age and length information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name.}

\item{time_series}{TRUE or FALSE}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a ggplot object
}
\description{
Function to plot length frequency of fish species
}
\examples{
\dontrun{
data(nwfsc_bio)
data(afsc_bio)
data(pbs_bio)
all_data <- bind_rows(nwfsc_bio, afsc_bio, pbs_bio)

length_frequency(all_data, species = "arrowtooth flounder")
length_frequency(all_data, c("NWFSC", "AK GULF"),species = "anoplopoma fimbria", facet_all = F)
}
}
