% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{length_frequency}
\alias{length_frequency}
\title{Function to plot length frequency of fish species}
\usage{
length_frequency(data)
}
\arguments{
\item{data}{a data frame made using pull_bio() from a fisheries survey, contains biological info}
}
\value{
a ggplot object
}
\description{
Function to plot length frequency of fish species
}
\examples{
\dontrun{

}
}
