% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{length_frequency}
\alias{length_frequency}
\title{Function to plot length frequency of fish species}
\usage{
length_frequency(data, time_series = FALSE)
}
\arguments{
\item{data}{a cleaned data frame from fisheries surveys, contains biological info}

\item{time_series}{TRUE or FALSE}
}
\value{
a ggplot object
}
\description{
Function to plot length frequency of fish species
}
\examples{
\dontrun{
# US West Coast
load("data/nwfsc_bio.rda")
data <- nwfsc_bio |> filter(common_name == "arrowtooth flounder")
length_frequency(data)

# Canada 
load("data/pbs_bio.rda")
data <- pbs_bio |> filter(common_name == "dover sole")
length_frequency(data, time_series = TRUE)

# Alaska
load("data/afsc_bio.rda")
data <- afsc_bio |> filter(common_name == "rex sole")
length_frequency(data, time_series = TRUE)

}
}
