% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_frequency.R
\name{length_frequency}
\alias{length_frequency}
\title{Function to plot length frequency of fish species}
\usage{
length_frequency(data, time_series = FALSE)
}
\arguments{
\item{data}{a data frame made using pull_bio() from a fisheries survey, contains biological info}

\item{time_series}{TRUE or FALSE}
}
\value{
a ggplot object
}
\description{
Function to plot length frequency of fish species
}
\examples{
\dontrun{
data <- pull_bio(survey = "NWFSC.Combo", common_name = "Pacific hake")
length_frequency(data)

data <- pull_bio(survey = "NWFSC.Combo", common_name = "Dover sole")
length_frequency(data, time_series = TRUE)
}
}
