% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_ts.R
\name{length_ts}
\alias{length_ts}
\title{Main function to plot mean lengths over time by sex}
\usage{
length_ts(data, species)
}
\arguments{
\item{data}{region bio data to be plotted. Preload from data file, see examples.}

\item{species}{common or scientific name of target species}
}
\value{
a time series plot of mean lengths by sex
}
\description{
Main function to plot mean lengths over time by sex
}
\examples{
\dontrun{
data("nwfsc_bio")
data("pbs_bio")
data("afsc_bio")
akgulf_bio <- afsc_bio \%>\% filter(survey == "AK GULF")
akbsai_bio <- afsc_bio \%>\% filter(survey == "AK BSAI")
all_data <- bind_rows(akgulf_bio, akbsai_bio, nwfsc_bio, pbs_bio)

length_ts(all_data, "arrowtooth flounder")
length_ts(akgulf_bio, "atka mackerel")
}
}
