% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{length_weight}
\alias{length_weight}
\title{Main function to plot length-weight relationships by sex}
\usage{
length_weight(data, species, subset = TRUE)
}
\arguments{
\item{data}{region bio data to be plotted. Preload from data file, see examples.}

\item{species}{common or scientific name of target species}

\item{subset}{default TRUE for a faster plotting subset of n = 2000. Set FALSE for all available data.}
}
\value{
a plot of sexed data with log regression slope and intercept
}
\description{
Main function to plot length-weight relationships by sex
}
\examples{
\dontrun{
data("nwfsc_bio")
data("pbs_bio")
data("afsc_bio")
akgulf_bio <- afsc_bio \%>\% filter(survey == "AK GULF")
akbsai_bio <- afsc_bio \%>\% filter(survey == "AK BSAI")
all_data <- bind_rows(akgulf_bio, akbsai_bio, nwfsc_bio, pbs_bio)

length_weight(all_data, "arrowtooth flounder")
length_weight(akgulf_bio, "atka mackerel", subset = FALSE)
}
}
