% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{length_weight}
\alias{length_weight}
\title{Main function to plot length-weight relationships by sex}
\usage{
length_weight(data, color = FALSE)
}
\arguments{
\item{data}{a data frame of target species including Length_cm, Weight_kg, and Sex}

\item{color}{default FALSE for greyscale, TRUE gives red and blue}
}
\value{
a plot of sexed data with log regression slope and intercept
}
\description{
Main function to plot length-weight relationships by sex
}
\examples{
\dontrun{
length_weight(bio.data)
}
}
