% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_vb_fit_doc.R
\docType{data}
\name{lw_predictions}
\alias{lw_predictions}
\title{lw_vb_fit_doc: Documentation of lw and vb predictions}
\format{
A data frame with 8444 rows and 9 variables:
\describe{
\item{fit_length}{fish lengths}
\item{fit_weight}{predicted weight at length}
\item{sex}{sex, M or F}
\item{a}{Scaling coefficient, intercept of length-weight regression }
\item{b}{exponent, slope of length-weight regression }
\item{region}{science center of data collection}
\item{common_name}{common species name}
\item{survey}{survey within region}
\item{scientific_name}{scientific species name}
}
}
\usage{
lw_predictions
}
\description{
Contains log-log length-weight regression estimates for the top species of each survey group, separated by sex.
}
\keyword{datasets}
