% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cpue.R
\name{map_cpue}
\alias{map_cpue}
\title{Function for empirical CPUE map for NOAA survey data}
\usage{
map_cpue(data, type = "static")
}
\arguments{
\item{data}{catch data from get_data() function from surveyjoin or pull_catch() from nwfscSurvey}

\item{type}{choose either "static" or "dynamic"}
}
\value{
a ggplot object
}
\description{
Function for empirical CPUE map for NOAA survey data
}
\examples{
\dontrun{
catch1 <- get_data(regions = "afsc", common = "sablefish")
map_cpue(catch1)

catch2 <- get_data(regions = "pbs", common = "arrowtooth flounder")
map_cpue(catch2, type = "dynamic")

catch3 <- pull_catch(survey = "NWFSC.Combo", common_name = "arrowtooth flounder")
map_cpue(catch3)
}
}
