% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data_document.R
\docType{data}
\name{pbs_bio}
\alias{pbs_bio}
\title{pbs_bio: Biological data for PBS survey}
\format{
A data frame with 226600 rows and 9 variables:
\describe{
\item{region}{region of data collection}
\item{survey}{subregion}
\item{year}{year specimen was collected}
\item{common_name}{common name of specien}
\item{scientific_name}{scientific name of specimen}
\item{sex}{specimen sex, M, F, or U (unknown)}
\item{length_cm}{length of specimen, in cm}
\item{weight_kg}{weight of specimen, in kg}
\item{age_years}{age of specimen, in years}
}
}
\usage{
pbs_bio
}
\description{
Contains biological data from PBS surveys for availible top species defined in data-raw/pbs_top.rds.
}
\keyword{datasets}
