% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_depth.R
\name{plot_age_depth}
\alias{plot_age_depth}
\title{Function for age-depth plot}
\usage{
plot_age_depth(
  data,
  subregions = c("NWFSC", "PBS", "AK BSAI", "AK GULF"),
  species,
  by_sex = FALSE,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing age and depth information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name.}

\item{by_sex}{show sex differentiation}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a ggplot object
}
\description{
Function for age-depth plot
}
\examples{
\dontrun{
data(afsc_bio)
data(pbs_bio)
data(nwfsc_bio)
all_data <- bind_rows(afsc_bio, pbs_bio, nwfsc_bio)

plot_age_depth(all_data, species = "arrowtooth flounder")
plot_age_depth(all_data, c("NWFSC", "AK GULF"),
 species = "anoplopoma fimbria",
 by_sex = T, facet_all = F)
}
}
