% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbi.R
\name{plot_dbi}
\alias{plot_dbi}
\title{Function and formatted data to plot design based index}
\usage{
plot_dbi(dbi_data, species, subregion)
}
\arguments{
\item{species}{common name of species of interest. See unique(data$common_name) for options per dataset.}

\item{subregion}{subregion required to specify subregion to be plotted. See unique(data$region) for options per dataset.}

\item{data}{survey data options loaded in script: nwfsc_biomass, afsc_biomass, pbs_biomass}
}
\value{
a ggplot object, plot of design based indicies
}
\description{
Function and formatted data to plot design based index
}
\examples{
\dontrun{
plot_dbi(nwfsc_biomass, "sablefish", "U.S. West Coast")
plot_dbi(afsc_biomass, "sablefish", "U.S. Gulf of Alaska")
plot_dbi(pbs_biomass, "sablefish", "SYN QCS")
}
}
