% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbi.R
\name{plot_dbi}
\alias{plot_dbi}
\title{Function and formatted data to plot design based index}
\usage{
plot_dbi(species, subsurvey)
}
\arguments{
\item{species}{common or scientific name of species of interest. See examples for options per dataset.}

\item{subsurvey}{subsurvey or survey group, required to specify servey/region to be plotted. See examples for options per dataset.}
}
\value{
a ggplot object, plot of design based indicies
}
\description{
Function and formatted data to plot design based index
}
\examples{
\dontrun{
# To see options for species and surveys in regions: "NWFSC", "AFSC", "PBS"
data(all.dbi)

all.dbi \%>\%
filter(region == "AFSC") \%>\%
 distinct(common_name)

all.dbi \%>\%
filter(region == "AFSC") \%>\%
distinct(survey) # do not use "U.S. Eastern Bering Sea Standard Region" ; use "U.S. Eastern Bering Sea Standard Plus NW Region"


# usage examples
plot_dbi("sablefish", "U.S. West Coast")
plot_dbi("dover sole", "U.S. Gulf of Alaska")
plot_dbi("arrowtooth flounder", "SYN QCS")
}
}
