% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{plot von Bertalanffy function predictions}
\usage{
plot_growth(data, predictions, region, species)
}
\arguments{
\item{data}{biological data containing age, length, and sex information}

\item{predictions}{von Bertalanffy prediction data in /data folder}

\item{region}{region AFSC, NWFSC, or PBS}

\item{species}{species common name}
}
\value{
a ggplot object
}
\description{
plot von Bertalanffy function predictions
}
\examples{
\dontrun{
data(vb_predictions)
data(pbs_bio)
data(afsc_bio)
data(nwfsc_bio)
data <- rbind(pbs_bio, afsc_bio, nwfsc_bio)

plot_growth(data, vb_predictions, "AFSC", "walleye pollock")
plot_growth(data, vb_predictions, region = c("AFSC", "NWFSC", "PBS"), "arrowtooth flounder")
}
}
