% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{plot von Bertalanffy function predictions}
\usage{
plot_growth(
  data,
  subregions = c("AK BSAI", "AK GULF", "NWFSC", "PBS"),
  species,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing age, length, and sex information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name .}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted.}
}
\value{
a ggplot object
}
\description{
plot von Bertalanffy function predictions
}
\examples{
\dontrun{
data(pbs_bio)
data(afsc_bio)
data(nwfsc_bio)
all_data <- bind_rows(pbs_bio, afsc_bio, nwfsc_bio)

plot_growth(all_data, species = "arrowtooth flounder")
plot_growth(all_data, c("NWFSC", "AK GULF"),species = "anoplopoma fimbria", facet_all = F)
}
}
