% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_dbi.R
\name{plot_stan_dbi}
\alias{plot_stan_dbi}
\title{Function and formatted data to plot standardized design based index}
\usage{
plot_stan_dbi(species, surveys)
}
\arguments{
\item{species}{common or scientific name of species of interest.}

\item{surveys}{region or surveys to be plotted. Takes a character list of survey names OR a single region ( "NWFSC", "AFSC", or "PBS").}
}
\value{
a ggplot object, plot of standardized design based indicies of 1+ surveys.
}
\description{
Function and formatted data to plot standardized design based index
}
\examples{
\dontrun{
plot_stan_dbi("sablefish", "PBS")
plot_stan_dbi("sablefish", c("U.S. West Coast", "U.S. Gulf of Alaska"))
}
}
