% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_table.R
\name{survey_table}
\alias{survey_table}
\title{Main function to display survey specimen counts}
\usage{
survey_table(
  data,
  subregions = c("AK BSAI", "AK GULF", "NWFSC", "PBS"),
  species,
  form = 2,
  facet_all = TRUE
)
}
\arguments{
\item{data}{biological data containing length and depth information for at least regions specified in \code{subregions}.}

\item{subregions}{choose NWFSC, PBS, AK GULF, and/or AK BSAI. Default all.}

\item{species}{species common or scientific name.}

\item{form}{choose 1 for tibble or 2 for ggplot}

\item{facet_all}{if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted}
}
\value{
a tibble or ggplot object
}
\description{
Main function to display survey specimen counts
}
\examples{
\dontrun{
data(nwfsc_bio)
data(afsc_bio)
data(pbs_bio)
all_data <- rbind(nwfsc_bio, afsc_bio, pbs_bio)

survey_table(all_data, species = "arrowtooth flounder", form = 1)
survey_table(all_data, species = "arrowtooth flounder", form = 2)
survey_table(all_data, c("NWFSC", "AK GULF"), species = "anoplopoma fimbria", facet_all = F)

}
}
