% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_table.R
\name{survey_table}
\alias{survey_table}
\title{Main function to display survey specimen counts}
\usage{
survey_table(data, form = c(1, 2))
}
\arguments{
\item{data}{a data frame of target species including Length_cm, Weight_kg, Age_years, Otosag_id, and Year.}

\item{form}{choose 1 for tibble or 2 for ggplot}
}
\value{
a tibble or ggplot object
}
\description{
Main function to display survey specimen counts
}
\examples{
\dontrun{
survey_table(bio.data, form = 2)
}
}
