% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_table.R
\name{survey_table}
\alias{survey_table}
\title{Main function to display survey specimen counts}
\usage{
survey_table(data, species, form = 2)
}
\arguments{
\item{data}{a data frame of bio data}

\item{species}{species of interest common or scientific name}

\item{form}{choose 1 for tibble or 2 for ggplot}
}
\value{
a tibble or ggplot object
}
\description{
Main function to display survey specimen counts
}
\examples{
\dontrun{
data("nwfsc_bio")
data("pbs_bio")
data("afsc_bio")
akgulf_bio <- afsc_bio \%>\% filter(survey == "AK GULF")
akbsai_bio <- afsc_bio \%>\% filter(survey == "AK BSAI")
all_data <- bind_rows(akgulf_bio, akbsai_bio, nwfsc_bio, pbs_bio)


survey_table(akgulf_bio, "atka mackerel", form = 1)
survey_table(nwfsc_bio, "sebastes zacentrus", form = 2)
survey_table(pbs_bio,"arrowtooth flounder", form = 2)
survey_table(all_data, "arrowtooth flounder")
}
}
