% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonb_growth.R
\name{vonb_growth}
\alias{vonb_growth}
\title{von Bertalanffy growth function plot}
\usage{
vonb_growth(data)
}
\arguments{
\item{data}{catch data from pull_bio(), containing Length_cm, Age_years, and Sex}
}
\value{
a ggplot object
}
\description{
von Bertalanffy growth function plot
}
\examples{
\dontrun{
US West Coast
load("data/nwfsc_bio.rda")
data <- nwfsc_bio |> filter(common_name == "arrowtooth flounder")
vonb_growth(data)

Canada
load("data/pbs_bio.rda")
data <- pbs_bio |> filter(common_name == "arrowtooth flounder")
vonb_growth(data)

Alaska
load("data/afsc_bio.rda")
data <- afsc_bio |> filter(common_name == "arrowtooth flounder")
vonb_growth(data)
}
}
